pyproject.toml
[tool.poetry]
name = "ppt"
version = "0.1.0"
description = ""
authors = ["Ana Clara Arantes <anaarantes@usp.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.2"
requests = "^2.31.0"
geemap = "^0.32.0"
ipykernel = "^6.29.4"
jupyterlab = "^4.1.6"
setuptools = "^69.5.1"
earthengine-api = "^0.1.399"
pyarrow = "^15.0.2"
fastparquet = "^2024.2.0"
rasterio = "^1.3.10"
pyproj = "^3.6.1"
ecmwf-api-client = "^1.6.3"
sqlalchemy = "^2.0.29"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
